mem_utils:otp: fix compiler warning
Fix the following warning messages. In the new OpenBMC, these warning
messages will become error messages.

```
aspeed-app/mem_utils/md.c:59:19: error: assignment to 'long unsigned
int' from 'void *' makes integer from pointer without a cast
[-Wint-conversion]
   59 |         virt_addr = map_base + (phy_addr & MAP_MASK);

aspeed-app/mem_utils/md.c:75:5: error: implicit declaration of function
'close'; did you mean 'pclose'? [-Wimplicit-function-declaration]

aspeed-app/otp/otp_ast2600.c:493:19: error: assignment to 'uint8_t *'
{aka 'unsigned char *'} from incompatible pointer type 'uint32_t *' {aka
'unsigned int *'} [-Wincompatible-pointer-types]
  493 |         xfer.data = data;

Fix error: ‘sb_sha224’ defined but not used.
```

+++ mem_utils/md.c
@@ -56,7 +56,7 @@ int main(int argc, char *argv[])
 
 /*	printf("mapping base = %08x , addr = %08x \n", map_base, addr); 
 	addr = map_base + (addr & MAP_MASK); */
-	virt_addr = map_base + (phy_addr & MAP_MASK); 
+	virt_addr = (unsigned long)map_base + (phy_addr & MAP_MASK);
 	
 	/* If another parameter, it is the length to display.
 	* Length is the number of objects, not number of bytes.
